Evitar eliminacion si hay factura asociada

+++ controllers/clienteController.js
@@ -1,4 +1,5 @@
 import Cliente from "../models/Cliente.js";
+import Factura from "../models/Factura.js";
 
 const nuevoCliente = async (req, res) => {
     const cliente = new Cliente(req.body);
@@ -6,18 +7,17 @@ const nuevoCliente = async (req, res) => {
     const { cedula } = req.body; //extraemos cedula.
     const existeCliente = await Cliente.findOne({ cedula });
 
-
     if (existeCliente) {
         const error = new Error("Cliente ya registrado");
         return res.status(400).json({ msg: error.message });
     }
     try {
         const clienteAlmacenado = await cliente.save();
-        res.json({msg:'Se ha almacenado el cliente correctamente'});
+        res.json({ msg: "Se ha almacenado el cliente correctamente" });
     } catch (e) {
         //console.log(error);
-        const error = new Error("Hubo un error inesperado al crear el cliente");//Mensaje de error
-        res.status(500).json({msg: error.message})
+        const error = new Error("Hubo un error inesperado al crear el cliente"); //Mensaje de error
+        res.status(500).json({ msg: error.message });
     }
 };
 
@@ -86,14 +86,25 @@ const eliminarCliente = async (req, res) => {
         const error = new Error("ID de cliente inválido");
         return res.status(404).json({ msg: error.message });
     }
+
     if (!cliente) {
         const error = new Error("Cliente no encontrado");
         return res.status(404).json({ msg: error.message });
     }
 
+    // Verificar si el cliente está presente en alguna factura
+    const factura = await Factura.findOne({ cliente: id });
+
+    if (factura) {
+        const error = new Error(
+            "No se puede eliminar el cliente, está asociado a una factura"
+        );
+        return res.status(400).json({ msg: error.message });
+    }
+
     try {
         await cliente.deleteOne();
-        res.json({ msg: "Cliente Eliminado" });
+        res.json({ msg: "Cliente eliminado" });
     } catch (error) {
         console.log(error);
     }
